remove trace and trace filter to independent crate
