+++ crates/client/src/rpc/types/cfx/mod.rs
@@ -1,10 +1,8 @@
 mod access_list;
 pub mod account;
-pub mod address;
 pub mod blame_info;
 pub mod block;
 pub mod consensus_graph_states;
-pub mod epoch_number;
 mod fee_history;
 pub mod filter;
 pub mod log;
@@ -17,6 +15,7 @@ pub mod status;
 pub mod storage_collateral_info;
 pub mod sync_graph_states;
 pub mod token_supply_info;
+pub mod trace;
 pub mod transaction;
 pub mod transaction_request;
 pub mod tx_pool;
@@ -24,9 +23,13 @@ pub mod vote_params_info;
 
 pub use access_list::*;
 pub use account::Account;
-pub use address::{
-    check_rpc_address_network, check_two_rpc_address_network_match,
-    RcpAddressNetworkInconsistent, RpcAddress, UnexpectedRpcAddressNetwork,
+pub use cfx_rpc_primitives::{
+    address,
+    address::{
+        check_rpc_address_network, check_two_rpc_address_network_match,
+        RcpAddressNetworkInconsistent, RpcAddress, UnexpectedRpcAddressNetwork,
+    },
+    epoch_number,
 };
 pub use fee_history::CfxFeeHistory;
 pub use sponsor_info::SponsorInfo;
