+++ crates/client/src/rpc/types/eth/mod.rs
@@ -2,16 +2,11 @@
 // Conflux is free software and distributed under GNU General Public License.
 // See http://www.gnu.org/licenses/
 
-mod trace;
-mod trace_filter;
-
 pub use cfx_rpc_eth_types::{
-    eth_pubsub, AccountPendingTransactions, Block, BlockNumber,
-    EthRpcLogFilter, FilterChanges, Header, Log, Receipt, SyncInfo, SyncStatus,
-    Transaction, TransactionRequest,
-};
-
-pub use self::{
+    eth_pubsub,
     trace::{LocalizedTrace, Res},
     trace_filter::TraceFilter,
+    AccountPendingTransactions, Block, BlockNumber, EthRpcLogFilter,
+    FilterChanges, Header, Log, Receipt, SyncInfo, SyncStatus, Transaction,
+    TransactionRequest,
 };
