+++ crates/rpc/rpc-cfx-types/src/lib.rs
@@ -1,0 +1,2 @@
+pub mod trace;
+pub mod trace_filter;
