+++ crates/rpc/rpc-cfx-types/src/trace.rs
@@ -2,8 +2,6 @@
 // Conflux is free software and distributed under GNU General Public License.
 // See http://www.gnu.org/licenses/
 
-use super::{eth::LocalizedTrace as EthLocalizedTrace, RpcAddress};
-use crate::rpc::types::Bytes;
 use cfx_addr::Network;
 use cfx_execute_helper::exec_tracer::{
     Action as VmAction, ActionType as VmActionType, BlockExecTraces,
@@ -12,13 +10,12 @@ use cfx_execute_helper::exec_tracer::{
     InternalTransferAction as VmInternalTransferAction,
     LocalizedTrace as PrimitiveLocalizedTrace, Outcome, TransactionExecTraces,
 };
-use cfx_executor::internal_contract::evm_map;
-use cfx_parameters::internal_contract_addresses::CROSS_SPACE_CONTRACT_ADDRESS;
-use cfx_types::{Space, H160, H256, U256, U64};
+use cfx_rpc_primitives::{Bytes, RpcAddress};
+use cfx_types::{Space, H256, U256, U64};
 use cfx_vm_types::{CallType, CreateType};
 use primitives::SignedTransaction;
 use serde::{ser::SerializeStruct, Deserialize, Serialize, Serializer};
-use std::{collections::HashMap, sync::Arc};
+use std::sync::Arc;
 use strum_macros::EnumDiscriminants;
 
 #[derive(Debug, Clone, PartialEq, EnumDiscriminants)]
@@ -346,7 +343,7 @@ impl LocalizedBlockTrace {
     ) -> Result<Self, String> {
         let traces: Vec<TransactionExecTraces> = traces.into();
         if traces.len() != transactions.len() {
-            bail!("trace and tx hash list length unmatch!");
+            error_chain::bail!("trace and tx hash list length unmatch!");
         }
         let transaction_traces = traces
             .into_iter()
@@ -370,33 +367,33 @@ impl LocalizedBlockTrace {
     }
 }
 
-#[derive(Debug, Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct EpochTrace {
-    cfx_traces: Vec<LocalizedTrace>,
-    eth_traces: Vec<EthLocalizedTrace>,
-    mirror_address_map: HashMap<H160, RpcAddress>,
-}
+// #[derive(Debug, Serialize)]
+// #[serde(rename_all = "camelCase")]
+// pub struct EpochTrace {
+//     cfx_traces: Vec<LocalizedTrace>,
+//     eth_traces: Vec<EthLocalizedTrace>,
+//     mirror_address_map: HashMap<H160, RpcAddress>,
+// }
 
-impl EpochTrace {
-    pub fn new(
-        cfx_traces: Vec<LocalizedTrace>, eth_traces: Vec<EthLocalizedTrace>,
-    ) -> Self {
-        let mut mirror_address_map = HashMap::new();
-        for t in &cfx_traces {
-            if let Action::Call(action) = &t.action {
-                if action.to.hex_address == CROSS_SPACE_CONTRACT_ADDRESS {
-                    mirror_address_map.insert(
-                        evm_map(action.from.hex_address).address,
-                        action.from.clone(),
-                    );
-                }
-            }
-        }
-        Self {
-            cfx_traces,
-            eth_traces,
-            mirror_address_map,
-        }
-    }
-}
+// impl EpochTrace {
+//     pub fn new(
+//         cfx_traces: Vec<LocalizedTrace>, eth_traces: Vec<EthLocalizedTrace>,
+//     ) -> Self {
+//         let mut mirror_address_map = HashMap::new();
+//         for t in &cfx_traces {
+//             if let Action::Call(action) = &t.action {
+//                 if action.to.hex_address == CROSS_SPACE_CONTRACT_ADDRESS {
+//                     mirror_address_map.insert(
+//                         evm_map(action.from.hex_address).address,
+//                         action.from.clone(),
+//                     );
+//                 }
+//             }
+//         }
+//         Self {
+//             cfx_traces,
+//             eth_traces,
+//             mirror_address_map,
+//         }
+//     }
+// }
