+++ crates/rpc/rpc-eth-types/Cargo.toml
@@ -1,0 +1,33 @@
+[package]
+name = "cfx-rpc-eth-types"
+edition = "2021"
+version.workspace = true
+authors.workspace = true
+description.workspace = true
+documentation.workspace = true
+homepage.workspace = true
+keywords.workspace = true
+repository.workspace = true
+license-file.workspace = true
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+rustc-hex = { workspace = true }
+serde = { workspace = true }
+serde_json = { workspace = true }
+cfx-types = { workspace = true }
+thiserror = { workspace = true }
+jsonrpc-core = "15.1.0"
+primitives = { workspace = true }
+cfxcore = { workspace = true }
+rlp = { workspace = true }
+cfx-parameters = { workspace = true }
+alloy-rpc-types = { workspace = true }
+serde-utils = { workspace = true }
+cfx-rpc-primitives = { workspace = true }
+cfx-execute-helper = { workspace = true }
+cfx-vm-types = { workspace = true }
+cfx-rpc-cfx-types = { workspace = true }
+error-chain = { workspace = true }
+
