+++ crates/rpc/rpc-eth-types/src/lib.rs
@@ -4,20 +4,21 @@ mod errors;
 pub mod eth_pubsub;
 mod filter;
 mod log;
-mod primitives;
 mod receipt;
 mod sync;
+pub mod trace;
+pub mod trace_filter;
 mod transaction;
 mod transaction_request;
 mod tx_pool;
 
 pub use block::{Block, Header};
 pub use block_number::BlockNumber;
+pub use cfx_rpc_primitives::{Bytes, U64};
 pub use errors::Error;
 pub use eth_pubsub::*;
 pub use filter::*;
 pub use log::Log;
-pub use primitives::{Bytes, U64};
 pub use receipt::Receipt;
 pub use sync::{SyncInfo, SyncStatus};
 pub use transaction::Transaction;
