+++ crates/rpc/rpc-eth-types/src/trace_filter.rs
@@ -16,11 +16,12 @@
 
 //! Trace filter deserialization.
 
-use crate::rpc::types::eth::BlockNumber;
+use crate::BlockNumber;
 use cfx_execute_helper::exec_tracer::TraceFilter as PrimitiveTraceFilter;
 use cfx_types::{Space, H160};
 use jsonrpc_core::Error as RpcError;
 use primitives::EpochNumber;
+use serde::Deserialize;
 use std::convert::TryInto;
 
 /// Trace filter
