--- crates/rpc/rpc_eth_types/src/primitives/mod.rs
@@ -1,5 +1,0 @@
-pub mod bytes;
-pub mod variadic_u64;
-
-pub use bytes::Bytes;
-pub use variadic_u64::U64;
